Refactor US class and improve ultrasonic sensor handling
- Changed pin data types in US class from uint8_t to int for better compatibility.
- Updated method signatures in US class to use int for pin parameters.
- Enhanced setupUS and getUSDistance functions for clearer distance reporting in millimeters.
- Refactored state machine logic for better readability and consistency.
- Improved servo control functions with better validation and initialization checks.
- Added detailed comments for clarity and maintainability.

+++ US.cpp
@@ -4,11 +4,11 @@
 US::US(){
 }
 
-US::US(uint8_t pinTrigger, uint8_t pinEcho){
+US::US(int pinTrigger, int pinEcho){
   US::init(pinTrigger,pinEcho);
 }
 
-void US::init(uint8_t pinTrigger, uint8_t pinEcho)
+void US::init(int pinTrigger, int pinEcho)
 {
   _pinTrigger = pinTrigger;
   _pinEcho = pinEcho;
